mode all hacks commands to hax.js
